+++ src/main/java/cn/edu/tsinghua/iotdb/kairosdb/conf/ConfigDescriptor.java
@@ -78,8 +78,8 @@ public class ConfigDescriptor {
         config.SPECIAL_TAG = properties.getProperty("SPECIAL_TAG", "SPECIAL_TAG");
         config.STORAGE_GROUP_SIZE = Integer
             .parseInt(properties.getProperty("STORAGE_GROUP_SIZE", "50"));
-        config.POINT_EDGE = Integer.parseInt(properties.getProperty("POINT_EDGE", "50000000"));
-        config.TIME_EDGE = Integer.parseInt(properties.getProperty("TIME_EDGE", "50000000"));
+        config.POINT_EDGE = Long.parseLong(properties.getProperty("POINT_EDGE", "50000000"));
+        config.TIME_EDGE = Long.parseLong(properties.getProperty("TIME_EDGE", "50000000"));
         config.MAX_ROLLUP = Integer
             .parseInt(properties.getProperty("MAX_ROLLUP", config.MAX_ROLLUP + ""));
         config.DEBUG = Integer.parseInt(properties.getProperty("DEBUG", config.DEBUG + ""));
